+++ js/data.js
@@ -1,0 +1,44 @@
+
+ export const id = {
+  1: '1',
+  2: '2',
+  3: '3',
+  4: '4',
+  5: '5',
+  6: '6',
+  7: '7',
+  8: '8',
+  9: '9',
+  10: '10',
+  11: '11',
+  12: '12',
+  13: '13',
+  14: '14',
+  15: '15',
+  16: '16',
+  17: '17',
+  18: '18',
+  19: '19',
+  20: '20',
+  21: '21',
+  21: '21',
+  22: '22',
+  23: '23',
+  24: '24',
+  25: '25',
+};
+
+
+
+export const description = {
+    op1: 'первое',
+    op2: 'второе',
+    op3: 'третье',
+    op4: 'четв',
+    op5: 'пятое',
+  };
+
+  export const nameFolov = ["Vasi", "Peti", "Andre"];
+  export  const max = 6;
+  export  const text = ["Всё отлично!", "В целом всё неплохо. Но не всё."];
+  export  const arrText = Math.floor(Math.random() * text.length);
